update: rename utils types file
