+++ src/auth/auth.middleware.ts
@@ -4,7 +4,7 @@ import { AuthService } from './auth.service';
 import { ApiKeyService } from '../utils/services/api.key.service.util';
 import { IpWhitelistService } from '../utils/services/ip.whitelist.service.util';
 import { User } from './auth.types';
-import { parseIp } from '../utils/common.util';
+import { parseIp } from '../utils/common/common.util';
 
 @Injectable()
 export class AuthMiddleware implements NestMiddleware {
