+++ src/main.ts
@@ -6,7 +6,7 @@ import { INestApplication, Logger } from '@nestjs/common';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { ExceptionsFilter } from './utils/filters/exception.filter.util';
 import { ApiResponseUtil } from './utils/common/api.response.util';
-import { Environment } from './utils/types.util';
+import { Environment } from './utils/utils.types';
 import { rateLimitMiddleware } from './utils/middlewares/rate-limit.middleware';
 import { RedisService } from './redis/redis.service';
 import { RedisCacheInterceptor } from './redis/redis.interceptor';
